SMS and Search

+++ src/main/java/VORTEX/DARIPROJECT/Repository/DepositRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import VORTEX.DARIPROJECT.ENTITY.Deposit_of_guarantees;
 
 @Repository
-public interface DepositRepo extends JpaRepository<Deposit_of_guarantees,Integer> {
+public interface DepositRepo extends JpaRepository<Deposit_of_guarantees,Long> {
  
 }
